Refactor codebase to improve type safety and readability
- Removed unnecessary static_cast<int>() conversions in various files to enhance clarity and prevent potential issues with signed/unsigned comparisons.
- Updated the implementation of AuthWorkerThreadPool to be fully inline in the header file, simplifying the code structure.
- Cleaned up SecureMemory implementation by moving all definitions inline in the header, improving maintainability.
- Fixed various instances of potential out-of-bounds access in CBOR parsing functions by removing redundant static_cast<int>() checks.
- Added PowerShell scripts for clang-tidy and LLVM syntax checking to streamline code quality checks in the development workflow.
- Improved error handling and logging in AuditBatchWriter and StreamWriter to ensure better diagnostics during runtime.
- Enhanced InputValidator to use a constant for maximum AQL query size, improving maintainability and readability.

+++ src/acceleration/cpu_backend_tbb.cpp
@@ -318,7 +318,7 @@ public:
                         }
                         
                         // Partial sort to get k nearest
-                        size_t actualK = std::min(k,static_cast<int>(distances.size()));
+                        size_t actualK = std::min(k, distances.size());
                         std::partial_sort(
                             distances.begin(),
                             distances.begin() + actualK,
