+++ src/acceleration/geo_acceleration_bridge.cpp
@@ -173,7 +173,7 @@ static int bridge_geo_containment(
     auto* geo = themis::geo::getGpuSpatialBackend();
     themis::geo::SpatialBatchResults res = geo->batchIntersects(batch);
     const size_t n = static_cast<size_t>(numPoints);
-    for (size_t i = 0; i <static_cast<int>(res.mask.size()) && i < n; ++i) {
+    for (size_t i = 0; i < res.mask.size() && i < n; ++i) {
         results[i] = res.mask[i];
     }
     return 0;
