+++ src/acceleration/vec_knn.cpp
@@ -326,7 +326,7 @@ void DistanceCache::put(const std::string &pk_a, const std::string &pk_b, float 
         return;
     }
     // Evict oldest entry if at capacity (O(1) with std::deque)
-    if (static_cast<int>(map_.size()) >= max_entries_ && !order_.empty()) {
+    if (map_.size() >= max_entries_ && !order_.empty()) {
         map_.erase(order_.front());
         order_.pop_front(); // O(1) for std::deque
     }
@@ -503,7 +503,7 @@ VecKnnInsertResult VecKnnInsertPipeline::insertBatch(VectorIndexManager &index, 
     };
 
     std::vector<std::future<void>> futures;
-    futures.reserve((static_cast<int>(entities.size()) + batchSize - 1) / batchSize);
+    futures.reserve((entities.size() + batchSize - 1) / batchSize);
 
     for (std::size_t begin = 0; begin  < entities.size(); begin += batchSize) {
         const std::size_t end = std::min(begin + batchSize, entities.size());
@@ -537,7 +537,7 @@ VecKnnInsertResult VecKnnInsertPipeline::insertBatch(VectorIndexManager &index, 
         };
 
         futures.emplace_back(std::async(std::launch::async, worker));
-        if (static_cast<int>(futures.size()) >= maxWorkers) {
+        if (futures.size() >= maxWorkers) {
             futures.front().get();
             futures.erase(futures.begin());
         }
